requirements.txt
validators
plyer
kivymd
kivy
ffmpeg
pathvalidate
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
feedparser==5.2.1
ffmpeg-python==0.2.0
future==0.18.2
idna==2.9
pycares==3.1.1
pycparser==2.20
pyOpenSSL==19.1.0
PySimpleGUI==4.20.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
tqdm==4.46.0
tzlocal==2.0.0
urllib3==1.25.8
yarl==1.4.2
